project/pyproject.toml
[tool.poetry]
name = "django_project_satvaspace"
version = "0.1.0"
description = ""
authors = ["odi1n"]

[tool.poetry.dependencies]
python = "^3.8"
psycopg2 = "^2.9.1"

factory-boy = "^3.2.0"
pytest-django = "^4.4.0"

Django = "^3.2.8"
django-environ = "^0.7.0"

djangorestframework = "^3.12.4"
Markdown = "^3.3.4"

redis = "^3.5.3"
celery = "^5.1.2"
django-model-utils = "^4.2.0"
django-admin-sortable2="^1.0.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
